Jump search algorithm

+++ 0x1E-search_algorithms/100-jump.c
@@ -30,7 +30,7 @@ int jump_search(int *array, size_t size, int value)
 			{
 				printf("Value checked array[%ld] = [%d]\n", temp, array[temp]);
 				if (array[temp] == value)
-					return (temp);
+					return ((int)temp);
 			}
 		}
 	}
